Launching mysqld server with an unknown option triggers the assertion in the process of its termination. While it should shutdown with clean error message.
This is because analysis of the server options and transfer of the state (SST) are run simultaneously, but mysqld does not provide any way for emergency completion of the SST script (which running in separate process). When the server finds an unknown option, it calls unireg_abort() function that can detect incomplete thread and can trigger assertion in debug mode.

First, we have incomplete thread, which created in order to interact with the SST process. Second, until the SST process yet not terminated, we have an incomplete thread, which waits on the "sst_cond_" conditional variable in the Galera code. However, this variable will be set only at the end of SST process, in the sst_received() callback function. Therefore, if the SST is continued, this thread also freezes.

To solve these problems, I added code that emergency completes the SST process by sending a SIGTERM signal to it when unireg_abort() was called. This leads to completion of the thread, which interacts with the SST process. Also, sst_received() callback function will be called after termination of the SST process, and it unlocks the thread, which waiting on the "sst_cond_" conditional variable (in the Galera code).

Also, I corrected a race condition in the Galera code - calling of the sst_received() callback immediately after sending SST request should not be terminated with an error, even if we have not yet shifted to the S_JOINING state.

+++ build-ps/build-binary.sh
@@ -62,6 +62,9 @@ TOKUDB_BACKUP_VERSION='@@TOKUDB_BACKUP_VERSION@@'
 TAR=${TAR:-tar}
 SCONS_ARGS=${SCONS_ARGS:-""}
 
+# keep build files
+KEEP_BUILD=0
+
 #-------------------------------------------------------------------------------
 #
 # Step-1: Set default configuration.
@@ -72,7 +75,7 @@ SCONS_ARGS=${SCONS_ARGS:-""}
 if ! getopt --test
 then
     go_out="$(getopt --options=iqGdvjt: \
-        --longoptions=i686,verbose,copygalera,debug,valgrind,with-jemalloc:,with-yassl,with-ssl:,tag: \
+        --longoptions=i686,verbose,copygalera,debug,valgrind,with-jemalloc:,with-yassl,keep-build,with-ssl:,tag: \
         --name="$(basename "$0")" -- "$@")"
     test $? -eq 0 || exit 1
     eval set -- $go_out
@@ -103,7 +106,7 @@ do
         CMAKE_OPTS="${CMAKE_OPTS:-} -DWITH_VALGRIND=ON"
         BUILD_COMMENT="${BUILD_COMMENT:-}-valgrind"
         ;;
-    -q | --verbose)
+    -q | --verbose )
         shift
         QUIET='VERBOSE=1'
         ;;
@@ -115,6 +118,10 @@ do
     --with-yassl )
         shift
         WITH_SSL_TYPE="bundled"
+        ;;
+    --keep-build )
+        shift
+        KEEP_BUILD=1
         ;;
     --with-ssl )
         shift
@@ -436,7 +443,10 @@ fi
     $TAR --owner=0 --group=0 -czf "$TARGETDIR/$PRODUCT_FULL_NAME.tar.gz" $PRODUCT_FULL_NAME
 ) || exit 1
 
-rm -rf $TARGETDIR/bld
-rm -rf $TARGETDIR/usr
+if [[ $KEEP_BUILD -eq 0 ]]
+then
+    rm -rf $TARGETDIR/bld
+    rm -rf $TARGETDIR/usr
+fi
 
 echo "Build Complete"
